Add file structure to the code (issue #9) (#13)
* Add file structure to the code (issue #9)
This adds a src directory which contains all of the fortran and python code. Within src, there is a model directory which will contain the fortran code and a processing directory, which will contain the python code for processing. The script run.py will be a wrapper script which runs both the model and the processing.

* Fixed type accidentally introduced into welcome.f90

* Use wrapper script to call function within model directory

Co-authored-by: Kristin Barton <knbarton93@gmail.com>